abstract class added, common methods substracted to the abstract class, added extensions
