+++ src/view/film-container-view.js
@@ -1,26 +1,13 @@
-import {createElement} from '../render.js';
+import AbstractView from './abstract-view.js';
 
 const createFilmContainerTemplate = () => `<section class="films-list">
       <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
       <div class="films-list__container"></div>
     </section>`;
 
-export default class FilmContainerView {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
+export default class FilmContainerView extends AbstractView {
 
   get template() {
     return createFilmContainerTemplate();
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
