+++ src/view/no-film-view.js
@@ -1,4 +1,4 @@
-import {createElement} from '../render.js';
+import AbstractView from './abstract-view.js';
 
 const createNoFilmTemplate = () => (
   `<section class="films-list">
@@ -6,22 +6,9 @@ const createNoFilmTemplate = () => (
     </section>`
 );
 
-export default class NoFilmView {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
+export default class NoFilmView extends AbstractView {
 
   get template() {
     return createNoFilmTemplate();
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
