+++ src/view/show-more-button-view.js
@@ -1,24 +1,11 @@
-import {createElement} from '../render.js';
+import AbstractView from './abstract-view.js';
 
 const createShowMoreButtonTemplate = () => `<button class="films-list__show-more">
   Show more</button>`;
 
-export default class ShowMoreButtonView {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
+export default class ShowMoreButtonView extends AbstractView {
 
   get template() {
     return createShowMoreButtonTemplate();
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
