+++ src/view/sort-list-view.js
@@ -1,4 +1,4 @@
-import {createElement} from '../render.js';
+import AbstractView from './abstract-view.js';
 
 const createSortListTemplate = () => `<ul class="sort">
             <li><a href="#" class="sort__button sort__button--active">Sort by default</a></li>
@@ -6,22 +6,9 @@ const createSortListTemplate = () => `<ul class="sort">
             <li><a href="#" class="sort__button">Sort by rating</a></li>
           </ul>`;
 
-export default class SortListView {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
+export default class SortListView extends AbstractView { 
 
   get template() {
     return createSortListTemplate();
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
